PhotoTransitioningUsingUIViewPropertyAnimator
This application demonstrates the usage of UIViewPropertyAnimator to create a fully interative and interruptible custom view controller transition. This application is a simple photo application that displays a grid of photos. Tap a photos to initiate a custom transition to into a one-up view of the photo. Once in one-up, drag down with one finger to interactively dismiss the one-up view controller. Put a finger down on the photo at any point during the transition to interrupt it.